Cosmetic renaming
Renamed function dependencies_check to check_deps.

Also renamed files in scripts/ to have a .sh suffix which are bash shell
snippits:

   dependencies_check -> check_deps.sh
   common -> common.sh

Note check_deps.sh already existed in the tree but was an empty file
probably intended for the function now contained within.

Also added a comment at the top of the .sh files which are sourced so
that syntax-highlighting editors which distinguish between different
/bin/sh syntaxes have a clue to use the bash variant.