Only track non-skipped stages in `PREV_` variables
Only set `PREV_STAGE`, `PREV_STAGE_DIR` and `PREV_ROOTFS_DIR` when the
stage is actually built.

When building a custom image, I wanted to make the least modifications
to this repo, so I've added `SKIP` files to `stage3` and `stage4`, and
created my own `stage5`, but this confused `copy_previous`, which was
trying to copy the rootfs from a un-built stage. This fixes this issue.