Update build-docker.sh for other platforms
The helper realpath does not exist on OS X.  Check for the implementation before using. Testing on Docker for Mac.  Docker for Mac requires that you increase the size of the Docker VM to match the disk space used in the build. I am using 256G.

Allow DOCKER to be set.  On systems without open docker commands, support DOCKER="sudo docker" bash build-docker.sh.
Testing this on ClearLinux.