Allow customizing exports (images) and document
There are use cases where pi-gen can be useful that need different
image generation that the current defaults.  Rather than force such
users to require carrying modified versions of the default exports,
allow users to specify alternate exports (e.g. instead of export-image
and export-noobs, or in addition to these).  We do this using a
mechanism similar to STAGE_LIST, except that the lists of images are
per-stage and default to paying attention to the presence of
EXPORT_IMAGE and EXPORT_NOOBS as is currently the case.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>