Use mkdosfs -I to be able to write to /dev/loop*
From mkdosfs' manual:
"It is typical for fixed disk devices to be partitioned so, by default,
you are not permitted to create a filesystem across the entire device.
mkfs.fat will complain and tell you that it refuses to work."

On some systems, mkdosfs identifies the loop device is an entire disk.
The -I flag makes mkdosfs work properly on such systems.