workflow: Build image with Docker 32 bits images
Building on 64 bits hosts cause issues with QEMU silently failing, which
would cause the image to be generated, but with some critical issues,
like CA certificates not working.

Building with 32 bits images on Docker solves this problem, because it
avoids the currently incompatibility with QEMU on 64 bits.

Also changing the branch to trigger compilation on push. This allows us
to trigger a build on GH runner by pushing to the force-push branch -
avoiding the wait for the daily build.

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>