Install raspbian repo from packagecloud
When moving to arm64, system is being configured as Debian instead of
Raspbian. That makes packagecloud script to configure a different repo,
which doesn't currently have our packages.

Since there is no functional benefit of using on dist or another, to
avoid copying all the packages (and doubling space used on packagecloud
account), just fixed it on raspbian.

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>