Fix keyboard configuration
keyboard-configuration scripts ignores debconf settings if it doesn't
detect an attached keyboard. Instead, it sets XKBOPTIONS to
lv3:ralt_switch, which is not a desired option. It also considers an
empty XKBOPTIONS string to be unset. This workaround sets it to a
placeholder value which is removed later. dpkg-reconfigure is run
so that debconf-get-selections prints the correct values.