Allow export image customisation
Add new variable EXPORT_CONFIG_DIR to set the location of the scripts
pigen will run when exporting an image. Setting this is optional. If
not specified, the current location is retained.

By utilising STAGE_LIST AND EXPORT_CONFIG_DIR, a user can construct
custom images out-of-tree without modification to any defaults.