Add examples directory
It may be useful for users to have some example pigen configurations
more aligned with their needs, rather than using (or hacking) the
default stages. The examples directory is intended to serve as a
starting point. It's not intended to be production code.

The min64 example is a minimal and generic 64-bit Debian system which
boots kernel8, runs systemd with basic wired networking, and has a
handful of core packages installed in a ~670MB rootfs.