mvp test03
Results:
Big success!

What worked at expected?
- Packages get installed & the script runs and works as expected (i.e. makes Activity folder and git clones)
- Image size is 1.1GB with sugar and lighDM which is expected and good.

What didn't work?
- Cannot configure lightdm
In stage3, after installing the packages in 00-packages (including lightdm and xserver-xorg), when the script ran it said that lightdm is not installed.
Should it be ran in chroot environment? or should it be ran in a substage of stage3?

Will try both and update in the next commit