Add bdebstrap and mmdebstrap example
This creates a minimal system using bdebstrap (which uses mmdebstrap)
to demonstrate how a layering approach could be used to replace stages.
Highly experimental.
It leverages the current multipart example to export an AB image after
mmdebstrap has completed execution.

Unlike debootstrap, mmdebstrap cleans up pretty well after it has
finished which makes it tricker for the simple scheme implemented
by multipart/export-image to estimate the on-disk size needed for the
root fs partition. As a result, it's much smaller compared to an image
generated by the min64 example and apt will likely run out of disk
space. But it's just experimental code at this point.